move: publish system plugin interface file
